+++ src/Contracts/Server/Server.php
@@ -82,9 +82,11 @@ interface Server
     public function authorizable(): bool;
 
     /**
-     * @param mixed $parameters
+     * Get a URL for the server.
+     *
+     * @param mixed|array $extra
      * @param bool|null $secure
      * @return string
      */
-    public function url($parameters = [], bool $secure = null): string;
+    public function url($extra = [], bool $secure = null): string;
 }