+++ src/Core/Server/ServerRepository.php
@@ -41,6 +41,11 @@ class ServerRepository implements RepositoryContract
     private ConfigRepository $config;
 
     /**
+     * @var array
+     */
+    private array $cache;
+
+    /**
      * ServerRepository constructor.
      *
      * @param IlluminateContainer $container
@@ -50,6 +55,7 @@ class ServerRepository implements RepositoryContract
     {
         $this->container = $container;
         $this->config = $config;
+        $this->cache = [];
     }
 
     /**
@@ -59,6 +65,10 @@ class ServerRepository implements RepositoryContract
     {
         if (empty($name)) {
             throw new InvalidArgumentException('Expecting a non-empty JSON:API server name.');
+        }
+
+        if (isset($this->cache[$name])) {
+            return $this->cache[$name];
         }
 
         $class = $this->config->get("jsonapi.servers.{$name}");
@@ -81,7 +91,7 @@ class ServerRepository implements RepositoryContract
         }
 
         if ($server instanceof ServerContract) {
-            return $server;
+            return $this->cache[$name] = $server;
         }
 
         throw new RuntimeException("Class for server {$name} is not a server instance.");